Fix various compiler warnings

+++ src/caliper/SnapshotRecord.cpp
@@ -90,11 +90,11 @@ SnapshotRecord::get(const Attribute& attr) const
         return Entry::empty;
 
     if (attr.store_as_value()) {
-        for (int i = 0; i < m_sizes.n_immediate; ++i)
+        for (size_t i = 0; i < m_sizes.n_immediate; ++i)
             if (m_attr_array[i] == attr.id())
                 return Entry(attr, m_data_array[i]);
     } else {
-        for (int i = 0; i < m_sizes.n_nodes; ++i)
+        for (size_t i = 0; i < m_sizes.n_nodes; ++i)
             for (cali::Node* node = m_node_array[i]; node; node = node->parent())
                 if (node->attribute() == attr.id())
                     return Entry(node);
@@ -122,13 +122,13 @@ SnapshotRecord::unpack(CaliperMetadataAccessInterface& db) const
     std::map< Attribute, std::vector<Variant> > rec;
 
     // unpack node entries 
-    for (int i = 0; i < m_sizes.n_nodes; ++i)
+    for (size_t i = 0; i < m_sizes.n_nodes; ++i)
         for (const cali::Node* node = m_node_array[i]; node; node = node->parent())
             if (node->attribute() != CALI_INV_ID)
                 rec[db.get_attribute(node->attribute())].push_back(node->data());
 
     // unpack immediate entries
-    for (int i = 0; i < m_sizes.n_immediate; ++i)
+    for (size_t i = 0; i < m_sizes.n_immediate; ++i)
         rec[db.get_attribute(m_attr_array[i])].push_back(m_data_array[i]);
 
     return rec;
